ASC-7 separated different parameters into a parameters package and added a config file to access them within calculator.py
